Rename CourseRosterService to RosterService
This makes room for follow up commits that will add assignment specific
methods.
